- Piotr Obrzut fixed orientation  getters (reported by Erwin Cumas)
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22850 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/physics/bullet/bullet.h
@@ -120,7 +120,7 @@ public:
   csRef<iDynamicsSystemCollider> GetCollider (unsigned int index) 
   {return colliders[index];}
   int GetColliderCount () 
-  {return colliders.GetSize ();}
+  {return (int) colliders.GetSize ();}
 
   CcdPhysicsEnvironment *GetBulletSys () 
   {return bullet_sys;}
